fix logout

+++ resources/js/components/layouts/Navigation/index.js
@@ -5,31 +5,39 @@ import { toast } from "react-toastify";
 import Logo from "../../../assets/images/shop.png";
 import Search from "./Search";
 import axios from "axios";
-import { deleteCookie, getCookie } from "./../../utils/cookie";
+import {
+    deleteAllCookies,
+    deleteCookie,
+    getCookie,
+} from "./../../utils/cookie";
 
 const apiLogout = "http://127.0.0.1:8000/api/auth/logout";
 
-export default function Navigation({ userProfile }) {
+export default function Navigation({ userProfile, loginState }) {
     const [click, setClick] = useState(false);
     const closeMobileMenu = () => setClick(false);
 
     const logout = async () => {
-        const headers = {
-            "Content-type": "application/json",
-            Authorization: `Bearer ${getCookie("access_token")}`,
-        };
-        await axios
-            .post(apiLogout, { data: "mydata" }, { headers: headers })
-            .then((res) => {
-                toast.success("Đăng xuất thành công!");
-                localStorage.removeItem("auth");
-                deleteCookie("access_token");
-                window.location.reload();
-            })
-            .catch((error) => {
-                toast.error("Đăng xuất không thành công!");
-                console.error(error);
-            });
+        if (getCookie("access_token") != "") {
+            const headers = {
+                "Content-type": "application/json",
+                Authorization: `Bearer ${getCookie("access_token")}`,
+            };
+            await axios
+                .post(apiLogout, { data: "mydata" }, { headers: headers })
+                .then((res) => {
+                    toast.success("Đăng xuất thành công!");
+                    deleteCookie("access_token");
+                    // deleteAllCookies();
+                    window.location.href = `/`;
+                })
+                .catch((error) => {
+                    toast.error("Đăng xuất không thành công!");
+                    console.error(error);
+                });
+        } else {
+            window.location.href = `/`;
+        }
     };
 
     return (
@@ -56,7 +64,7 @@ export default function Navigation({ userProfile }) {
                         </a>
                     </div>
                 </div>
-                {!localStorage.getItem("auth") ? (
+                {!loginState ? (
                     <ul className="signin-up">
                         <li className="sign-in" onClick={closeMobileMenu}>
                             <a href="/login">
